test: refactore jobstop and listmigrationtarget CLI tests
